+++ udata/core/topic/models.py
@@ -1,5 +1,6 @@
+from datetime import datetime
 from flask import url_for
-
+from mongoengine.fields import DateTimeField
 from mongoengine.signals import pre_save
 from udata.models import db
 from udata.search import reindex
@@ -9,7 +10,7 @@ from udata.tasks import as_task_param
 __all__ = ('Topic', )
 
 
-class Topic(db.Document):
+class Topic(db.Document, db.Owned):
     name = db.StringField(required=True)
     slug = db.SlugField(max_length=255, required=True, populate_from='name',
                         update=True, follow=True)
@@ -19,14 +20,25 @@ class Topic(db.Document):
 
     tags = db.ListField(db.StringField())
     datasets = db.ListField(
-        db.ReferenceField('Dataset', reverse_delete_rule=db.PULL))
+        db.LazyReferenceField('Dataset', reverse_delete_rule=db.PULL))
     reuses = db.ListField(
-        db.ReferenceField('Reuse', reverse_delete_rule=db.PULL))
+        db.LazyReferenceField('Reuse', reverse_delete_rule=db.PULL))
 
-    owner = db.ReferenceField('User')
     featured = db.BooleanField()
     private = db.BooleanField()
     extras = db.ExtrasField()
+
+    created_at = DateTimeField(default=datetime.utcnow, required=True)
+
+    meta = {
+        'indexes': [
+            '$name',
+            'created_at',
+            'slug'
+        ] + db.Owned.meta['indexes'],
+        'ordering': ['-created_at'],
+        'auto_create_index_on_save': True
+    }
 
     def __str__(self):
         return self.name
@@ -42,11 +54,15 @@ class Topic(db.Document):
         except cls.DoesNotExist:
             datasets_list_dif = document.datasets
         for dataset in datasets_list_dif:
-            reindex.delay(*as_task_param(dataset))
+            reindex.delay(*as_task_param(dataset.fetch()))
 
     @property
     def display_url(self):
         return url_for('topics.display', topic=self)
+
+    def count_discussions(self):
+        # There are no metrics on Topic to store discussions count
+        pass
 
 
 pre_save.connect(Topic.pre_save, sender=Topic)