+++ udata/harvest/backends/dcat.py
@@ -1,9 +1,8 @@
 import logging
 
-import requests
-
-from rdflib import Graph, URIRef, BNode
+from rdflib import Graph, URIRef
 from rdflib.namespace import RDF
+import xml.etree.ElementTree as ET
 from typing import List
 
 from udata.rdf import (
@@ -34,6 +33,9 @@ KNOWN_PAGINATION = (
     (HYDRA.PartialCollectionView, HYDRA.next),
     (HYDRA.PagedCollection, HYDRA.nextPage)
 )
+
+CSW_NAMESPACE = 'http://www.opengis.net/cat/csw/2.0.2'
+OWS_NAMESPACE = 'http://www.opengis.net/ows'
 
 # Useful to patch essential failing URIs
 URIS_TO_REPLACE = {
@@ -66,7 +68,8 @@ class DcatBackend(BaseBackend):
         # if format can't be guessed from the url
         # we fallback on the declared Content-Type
         if not fmt:
-            response = requests.head(self.source.url)
+            response = self.head(self.source.url)
+            response.raise_for_status()
             mime_type = response.headers.get('Content-Type', '').split(';', 1)[0]
             if not mime_type:
                 msg = 'Unable to detect format from extension or mime type'
@@ -86,7 +89,9 @@ class DcatBackend(BaseBackend):
         page = 0
         while url:
             subgraph = Graph(namespace_manager=namespace_manager)
-            data = requests.get(url).text
+            response = self.get(url)
+            response.raise_for_status()
+            data = response.text
             for old_uri, new_uri in URIS_TO_REPLACE.items():
                 data = data.replace(old_uri, new_uri)
             subgraph.parse(data=data, format=fmt)
@@ -119,6 +124,8 @@ class DcatBackend(BaseBackend):
         raise ValueError(f'Unable to find dataset with DCT.identifier:{item.remote_id}')
 
     def process(self, item):
+        if item.remote_id == 'None':
+            raise ValueError('The DCT.identifier is missing on this DCAT.Dataset record')
         graph = Graph(namespace_manager=namespace_manager)
         data = self.job.data['graphs'][item.kwargs['page']]
         format = self.job.data['format']
@@ -129,3 +136,89 @@ class DcatBackend(BaseBackend):
         dataset = self.get_dataset(item.remote_id)
         dataset = dataset_from_rdf(graph, dataset, node=node)
         return dataset
+
+
+class CswDcatBackend(DcatBackend):
+    display_name = 'CSW-DCAT'
+
+    DCAT_SCHEMA = 'http://www.w3.org/ns/dcat#'
+
+    def parse_graph(self, url: str, fmt: str) -> List[Graph]:
+        body = '''<csw:GetRecords xmlns:csw="http://www.opengis.net/cat/csw/2.0.2"
+                                  xmlns:gmd="http://www.isotc211.org/2005/gmd"
+                                  service="CSW" version="2.0.2" resultType="results"
+                                  startPosition="{start}" maxPosition="200"
+                                  outputSchema="{schema}">
+                    <csw:Query typeNames="gmd:MD_Metadata">
+                        <csw:ElementSetName>full</csw:ElementSetName>
+                        <ogc:SortBy xmlns:ogc="http://www.opengis.net/ogc">
+                            <ogc:SortProperty>
+                                <ogc:PropertyName>identifier</ogc:PropertyName>
+                            <ogc:SortOrder>ASC</ogc:SortOrder>
+                            </ogc:SortProperty>
+                        </ogc:SortBy>
+                    </csw:Query>
+                </csw:GetRecords>'''
+        headers = {'Content-Type': 'application/xml'}
+
+        graphs = []
+        page = 0
+        start = 1
+        response = self.post(url, data=body.format(start=start, schema=self.DCAT_SCHEMA),
+                             headers=headers)
+        response.raise_for_status()
+        content = response.text
+        tree = ET.fromstring(content)
+        if tree.tag == '{' + OWS_NAMESPACE + '}ExceptionReport':
+            raise ValueError(f'Failed to query CSW:\n{content}')
+        while tree:
+            graph = Graph(namespace_manager=namespace_manager)
+            search_results = tree.find('csw:SearchResults', {'csw': CSW_NAMESPACE})
+            if not search_results:
+                log.error(f'No search results found for {url} on page {page}')
+                break
+            for child in search_results:
+                subgraph = Graph(namespace_manager=namespace_manager)
+                subgraph.parse(data=ET.tostring(child), format=fmt)
+                graph += subgraph
+
+                for node in subgraph.subjects(RDF.type, DCAT.Dataset):
+                    id = subgraph.value(node, DCT.identifier)
+                    kwargs = {'nid': str(node), 'page': page}
+                    kwargs['type'] = 'uriref' if isinstance(node, URIRef) else 'blank'
+                    self.add_item(id, **kwargs)
+            graphs.append(graph)
+            page += 1
+
+            next_record = int(search_results.attrib['nextRecord'])
+            matched_count = int(search_results.attrib['numberOfRecordsMatched'])
+            returned_count = int(search_results.attrib['numberOfRecordsReturned'])
+
+            # Break conditions copied gratefully from
+            # noqa https://github.com/geonetwork/core-geonetwork/blob/main/harvesters/src/main/java/org/fao/geonet/kernel/harvest/harvester/csw/Harvester.java#L338-L369
+            break_conditions = (
+                # standard CSW: A value of 0 means all records have been returned.
+                next_record == 0,
+
+                # Misbehaving CSW server returning a next record > matched count
+                next_record > matched_count,
+
+                # No results returned already
+                returned_count == 0,
+
+                # Current next record is lower than previous one
+                next_record < start,
+
+                # Enough items have been harvested already
+                self.max_items and len(self.job.items) >= self.max_items
+            )
+
+            if any(break_conditions):
+                break
+
+            start = next_record
+            tree = ET.fromstring(
+                self.post(url, data=body.format(start=start, schema=self.DCAT_SCHEMA),
+                          headers=headers).text)
+
+        return graphs