+++ udata/harvest/tests/test_dcat_backend.py
@@ -4,6 +4,7 @@ import os
 import pytest
 
 from datetime import date
+import xml.etree.ElementTree as ET
 
 from udata.models import Dataset
 from udata.core.organization.factories import OrganizationFactory
@@ -19,6 +20,7 @@ log = logging.getLogger(__name__)
 TEST_DOMAIN = 'data.test.org'  # Need to be used in fixture file
 DCAT_URL_PATTERN = 'http://{domain}/{path}'
 DCAT_FILES_DIR = os.path.join(os.path.dirname(__file__), 'dcat')
+CSW_DCAT_FILES_DIR = os.path.join(os.path.dirname(__file__), 'csw_dcat')
 
 
 def mock_dcat(rmock, filename, path=None):
@@ -40,6 +42,19 @@ def mock_pagination(rmock, path, pattern):
             return dcatfile.read()
 
     rmock.get(rmock.ANY, text=callback)
+    return url
+
+
+def mock_csw_pagination(rmock, path, pattern):
+    url = DCAT_URL_PATTERN.format(path=path, domain=TEST_DOMAIN)
+
+    def callback(request, context):
+        request_tree = ET.fromstring(request.body)
+        page = int(request_tree.get('startPosition'))
+        with open(os.path.join(CSW_DCAT_FILES_DIR, pattern.format(page))) as cswdcatfile:
+            return cswdcatfile.read()
+
+    rmock.post(rmock.ANY, text=callback)
     return url
 
 
@@ -281,21 +296,32 @@ class DcatBackendTest:
         assert dataset.temporal_coverage is not None
         assert dataset.temporal_coverage.start == date(2016, 1, 1)
         assert dataset.temporal_coverage.end == date(2016, 12, 5)
+        assert dataset.frequency is None
 
-        assert len(dataset.resources) == 2
+        assert len(dataset.resources) == 3
 
         resource_1 = next(res for res in dataset.resources if res.title == 'Resource 1-1')
+        assert resource_1.filetype == 'remote'
         # Format is a IANA URI
         assert resource_1.format == 'json'
         assert resource_1.mime == 'application/json'
         assert resource_1.filesize == 12323
         assert resource_1.description == 'A JSON resource'
         assert resource_1.url == 'http://data.test.org/datasets/1/resources/1/file.json'
+        assert resource_1.type == 'main'
 
         resource_2 = next(res for res in dataset.resources if res.title == 'Resource 1-2')
         assert resource_2.format == 'json'
         assert resource_2.description == 'A JSON resource'
         assert resource_2.url == 'http://data.test.org/datasets/1/resources/2/file.json'
+        assert resource_2.type == 'main'
+
+        # Make sure additionnal resource is correctly harvested
+        resource_3 = next(res for res in dataset.resources if res.title == 'Resource 1-3')
+        assert resource_3.format == 'json'
+        assert resource_3.description == ''
+        assert resource_3.url == 'http://data.test.org/datasets/1/resources/3'
+        assert resource_3.type == 'other'
 
     def test_geonetwork_xml_catalog(self, rmock):
         url = mock_dcat(rmock, 'geonetwork.xml', path='catalog.xml')
@@ -343,6 +369,19 @@ class DcatBackendTest:
         assert dataset.harvest.uri == 'https://sig.oreme.org/geonetwork/srv/eng/catalog.search#/metadata//datasets/0437a976-cff1-4fa6-807a-c23006df2f8f'  # noqa
         assert dataset.harvest.remote_url == 'https://sig.oreme.org/geonetwork/srv/eng/catalog.search#/metadata//datasets/0437a976-cff1-4fa6-807a-c23006df2f8f'  # noqa
         assert dataset.harvest.last_update.date() == date.today()
+
+    def test_user_agent_get(self, rmock):
+        url = mock_dcat(rmock, 'catalog.xml', path='without/extension')
+        rmock.head(url, headers={'Content-Type': 'application/xml; charset=utf-8'})
+        get_mock = rmock.get(url)
+        org = OrganizationFactory()
+        source = HarvestSourceFactory(backend='dcat',
+                                      url=url,
+                                      organization=org)
+        actions.run(source.slug)
+
+        assert 'User-Agent' in get_mock.last_request.headers
+        assert get_mock.last_request.headers['User-Agent'] == 'uData/0.1 dcat'
 
     def test_unsupported_mime_type(self, rmock):
         url = DCAT_URL_PATTERN.format(path='', domain=TEST_DOMAIN)
@@ -408,3 +447,81 @@ class DcatBackendTest:
         job = source.get_last_job()
         assert len(job.items) == 0
         assert job.status == 'done'
+
+    def test_target_404(self, rmock):
+        filename = 'obvious-format.jsonld'
+        url = DCAT_URL_PATTERN.format(path=filename, domain=TEST_DOMAIN)
+        rmock.get(url, status_code=404)
+
+        source = HarvestSourceFactory(backend='dcat', url=url, organization=OrganizationFactory())
+        actions.run(source.slug)
+        source.reload()
+
+        job = source.get_last_job()
+        assert job.status == "failed"
+        assert len(job.errors) == 1
+        assert "404 Client Error" in job.errors[0].message
+
+        filename = 'need-to-head-to-guess-format'
+        url = DCAT_URL_PATTERN.format(path=filename, domain=TEST_DOMAIN)
+        rmock.head(url, status_code=404)
+
+        source = HarvestSourceFactory(backend='dcat', url=url, organization=OrganizationFactory())
+        actions.run(source.slug)
+        source.reload()
+
+        job = source.get_last_job()
+        assert job.status == "failed"
+        assert len(job.errors) == 1
+        assert "404 Client Error" in job.errors[0].message
+
+
+@pytest.mark.usefixtures('clean_db')
+@pytest.mark.options(PLUGINS=['csw-dcat'])
+class CswDcatBackendTest:
+
+    def test_geonetworkv4(self, rmock):
+        url = mock_csw_pagination(rmock, 'geonetwork/srv/eng/csw.rdf', 'geonetworkv4-page-{}.xml')
+        org = OrganizationFactory()
+        source = HarvestSourceFactory(backend='csw-dcat',
+                                      url=url,
+                                      organization=org)
+
+        actions.run(source.slug)
+
+        source.reload()
+
+        job = source.get_last_job()
+        assert len(job.items) == 6
+
+        datasets = {d.harvest.dct_identifier: d for d in Dataset.objects}
+
+        assert len(datasets) == 6
+
+        # First dataset
+        dataset = datasets['https://www.geo2france.fr/2017/accidento']
+        assert dataset.title == 'Localisation des accidents de la circulation routière en 2017'
+        assert dataset.description == 'Accidents corporels de la circulation en Hauts de France (2017)'
+        assert set(dataset.tags) == set([
+            'donnee-ouverte', 'accidentologie', 'accident', 'reseaux-de-transport', 'accident-de-la-route',
+            'hauts-de-france', 'nord', 'pas-de-calais', 'oise', 'somme', 'aisne'
+        ])
+        assert dataset.harvest.created_at.date() == date(2017, 1, 1)
+        assert len(dataset.resources) == 1
+        resource = dataset.resources[0]
+        assert resource.title == 'accidento_hdf_L93'
+        assert resource.url == 'https://www.geo2france.fr/geoserver/cr_hdf/ows'
+        assert resource.format == 'ogc:wms'
+
+    def test_user_agent_post(self, rmock):
+        url = mock_csw_pagination(rmock, 'geonetwork/srv/eng/csw.rdf', 'geonetworkv4-page-{}.xml')
+        get_mock = rmock.post(url)
+        org = OrganizationFactory()
+        source = HarvestSourceFactory(backend='csw-dcat',
+                                      url=url,
+                                      organization=org)
+
+        actions.run(source.slug)
+
+        assert 'User-Agent' in get_mock.last_request.headers
+        assert get_mock.last_request.headers['User-Agent'] == 'uData/0.1 csw-dcat'