+++ udata/tests/api/test_datasets_api.py
@@ -21,6 +21,7 @@ from udata.core.dataset.models import (HarvestDatasetMetadata,
                                        HarvestResourceMetadata, ResourceMixin)
 from udata.core.organization.factories import OrganizationFactory
 from udata.core.spatial.factories import SpatialCoverageFactory
+from udata.core.topic.factories import TopicFactory
 from udata.core.user.factories import AdminFactory, UserFactory
 from udata.i18n import gettext as _
 from udata.models import (LEGACY_FREQUENCIES, RESOURCE_TYPES,
@@ -139,6 +140,8 @@ class DatasetAPITest(APITestCase):
         license_dataset = VisibleDatasetFactory(license=LicenseFactory(id='cc-by'))
         format_dataset = DatasetFactory(resources=[ResourceFactory(format='my-format')])
         featured_dataset = VisibleDatasetFactory(featured=True)
+        topic_dataset = VisibleDatasetFactory()
+        topic = TopicFactory(datasets=[topic_dataset])
 
         paca, _, _ = create_geozones_fixtures()
         geozone_dataset = VisibleDatasetFactory(spatial=SpatialCoverageFactory(zones=[paca.id]))
@@ -230,6 +233,21 @@ class DatasetAPITest(APITestCase):
         self.assert200(response)
         self.assertEqual(len(response.json['data']), 1)
         self.assertEqual(response.json['data'][0]['id'], str(schema_version2_dataset.id))
+
+        # filter on topic
+        response = self.get(url_for('api.datasets', topic=topic.id))
+        self.assert200(response)
+        self.assertEqual(len(response.json['data']), 1)
+        self.assertEqual(response.json['data'][0]['id'], str(topic_dataset.id))
+
+        # filter on non existing topic
+        response = self.get(url_for('api.datasets', topic=topic_dataset.id))
+        self.assert200(response)
+        self.assertTrue(len(response.json['data']) > 0)
+
+        # filter on non id for topic
+        response = self.get(url_for('api.datasets', topic='xxx'))
+        self.assert400(response)
 
     def test_dataset_api_get(self):
         '''It should fetch a dataset from the API'''
@@ -1323,6 +1341,21 @@ class CommunityResourceAPITest(APITestCase):
         data = json.loads(response.data)
         self.assertEqual(data['id'], str(community_resource.id))
 
+    def test_resources_api_list(self):
+        '''It should list community resources from the API'''
+        community_resources = [CommunityResourceFactory() for _ in range(40)]
+        response = self.get(url_for('api.community_resources'))
+        self.assert200(response)
+        resources = json.loads(response.data)['data']
+
+        response = self.get(url_for('api.community_resources', page=2))
+        self.assert200(response)
+        resources += json.loads(response.data)['data']
+
+        self.assertEqual(len(resources), len(community_resources))
+        # Assert we don't have duplicates
+        self.assertEqual(len(set(res['id'] for res in resources)), len(community_resources))
+
     def test_community_resource_api_get_from_string_id(self):
         '''It should fetch a community resource from the API'''
         community_resource = CommunityResourceFactory()
@@ -1824,14 +1857,15 @@ class HarvestMetadataAPITest:
         response = api.get(url_for('api.dataset', dataset=dataset))
         assert200(response)
         assert response.json['created_at'] == creation_date.isoformat()
-        assert response.json['last_modified'] != modification_date.isoformat()
+        assert response.json['last_modified'] == modification_date.isoformat()
 
         resource_harvest_metadata = HarvestResourceMetadata(
             created_at=creation_date,
+            modified_at=modification_date,
         )
         dataset = DatasetFactory(resources=[ResourceFactory(harvest=resource_harvest_metadata)])
 
         response = api.get(url_for('api.dataset', dataset=dataset))
         assert200(response)
         assert response.json['resources'][0]['created_at'] == creation_date.isoformat()
-        assert response.json['resources'][0]['last_modified'] != modification_date.isoformat()
+        assert response.json['resources'][0]['last_modified'] == modification_date.isoformat()